+++ lua/config/keymap.lua
@@ -43,9 +43,74 @@ map('n', '<leader>fr', function()
   pickers().resume()
 end, { desc = 'Resume picker' })
 
+map('n', '<leader>f.', function()
+  pickers().oldfiles()
+end, { desc = 'Recent files' })
+
+map('n', '<leader>fd', function()
+  pickers().document_symbols()
+end, { desc = 'Document symbols' })
+
+map('n', '<leader>fk', function()
+  pickers().keymaps()
+end, { desc = 'Search keymaps' })
+
 map('n', '<leader>fs', function()
   pickers().symbols()
 end, { desc = 'Search symbols' })
+
+map('n', '<leader>fo', '<cmd>AerialToggle!<cr>', { desc = 'Symbol outline' })
+
+map('n', '<leader>sd', function()
+  pickers().diagnostics()
+end, { desc = 'Diagnostics picker' })
+
+map('n', '<leader>sw', function()
+  pickers().grep_word()
+end, { desc = 'Grep word under cursor' })
+
+map('n', '<leader>xt', '<cmd>TodoTrouble<cr>', { desc = 'Project TODOs' })
+
+map('n', '<leader>gh', function()
+  pickers().git_hunks()
+end, { desc = 'Git hunks picker' })
+
+map('n', '<leader>cr', function()
+  pickers().lsp 'references'
+end, { desc = 'LSP references' })
+
+map('n', '<leader>ci', function()
+  pickers().lsp 'implementation'
+end, { desc = 'LSP implementation' })
+
+map('n', '<leader>ct', function()
+  pickers().lsp 'type_definition'
+end, { desc = 'LSP type definition' })
+
+map('n', '<leader>cl', function()
+  vim.wo.cursorline = not vim.wo.cursorline
+end, { desc = 'Toggle cursor line' })
+
+map('n', '<leader>y', function()
+  local path = vim.fn.expand '%:p'
+  if path == '' then
+    path = vim.fn.getcwd()
+  end
+  vim.fn.setreg('+', path)
+  require('custom.msgarea').echo_status('Yanked: ' .. path)
+end, { desc = 'Yank file path' })
+
+map('n', '<leader>qq', function()
+  local winid = vim.fn.win_findbuf(vim.fn.bufnr '[Location List]')
+  if winid and #winid > 0 then
+    vim.cmd.lclose()
+  else
+    vim.cmd.lopen()
+  end
+end, { desc = 'Toggle location list' })
+
+map('n', ']q', '<cmd>cnext<cr>', { desc = 'Next quickfix item' })
+map('n', '[q', '<cmd>cprev<cr>', { desc = 'Previous quickfix item' })
 
 map('n', '<leader>e', '<cmd>Oil<cr>', { desc = 'Edit filesystem' })
 map('n', '-', '<cmd>Oil<cr>', { desc = 'Oil' })
@@ -59,9 +124,10 @@ map('n', '<leader>sr', function()
   require('grug-far').open()
 end, { desc = 'Search and replace' })
 
-map('n', '<leader>fo', '<cmd>AerialToggle!<cr>', { desc = 'Symbol outline' })
-
 map('n', '<leader>vl', '<cmd>Lazy<cr>', { desc = 'Lazy' })
 map('n', '<leader>vm', '<cmd>Mason<cr>', { desc = 'Mason' })
+map('n', '<leader>vr', function()
+  require('custom.usage').report()
+end, { desc = 'Usage report' })
 
 require('custom.windows').setup()